+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Added missing conditional to reuse `exitFullscreen` method @rafa8626
 * Added new code snippet to use player with `AudioContext.decodeAudioData()` @rafa8626
+* Added missing conditional for bug with hidden iframes in Firefox @rafa8626
 
 *4.1.0 (2017/05/04)*